Move coverage and use applyreduce
