+++ src/GeometryOps.jl
@@ -28,6 +28,7 @@ include("methods/centroid.jl")
 include("methods/distance.jl")
 include("methods/equals.jl")
 include("methods/clipping/clipping_processor.jl")
+include("methods/clipping/coverage.jl")
 include("methods/clipping/cut.jl")
 include("methods/clipping/intersection.jl")
 include("methods/clipping/difference.jl")
